presentation of perm groups
