+++ sympy/__init__.py
@@ -80,7 +80,7 @@ from .plotting import plot, textplot, plot_backends, plot_implicit
 from .printing import pretty, pretty_print, pprint, pprint_use_unicode, \
     pprint_try_use_unicode, print_gtk, print_tree, pager_print, TableForm
 from .printing import rcode, ccode, fcode, jscode, julia_code, mathematica_code, \
-    octave_code, latex, preview, rust_code
+    octave_code, latex, preview, rust_code, mathml, cxxcode
 from .printing import python, print_python, srepr, sstr, sstrrepr
 from .interactive import init_session, init_printing
