+++ sympy/combinatorics/fp_groups.py
@@ -8,9 +8,11 @@ from sympy.printing.defaults import DefaultPrinting
 from sympy.utilities import public
 from sympy.utilities.iterables import flatten
 from sympy.combinatorics.free_groups import FreeGroupElement, free_group, zero_mul_simp
+from sympy.combinatorics.coset_table import (CosetTable,
+                                             coset_enumeration_r,
+                                             coset_enumeration_c)
 
-from itertools import chain, product
-from bisect import bisect_left
+from itertools import product
 
 
 @public
@@ -89,21 +91,32 @@ class FpGroup(DefaultPrinting):
         g = FpGroup(g[0].group, rels)
         return g
 
-    def coset_enumeration(self, H, strategy="relator_based", max_cosets=None):
+    def coset_enumeration(self, H, strategy="relator_based", max_cosets=None,
+                                                        resume=None, incomplete=False):
         """
         Return an instance of ``coset table``, when Todd-Coxeter algorithm is
         run over the ``self`` with ``H`` as subgroup, using ``strategy``
         argument as strategy. The returned coset table is compressed but not
         standardized.
 
+        An instance of `CosetTable` for `fp_grp` can be passed as the keyword
+        argument `resume` in which case the coset enumeration will start with
+        that instance and attempt to complete it.
+
+        When `incomplete` is `True` and the function is unable to complete for
+        some reason, the partially complete table will be returned.
+
         """
         if not max_cosets:
             max_cosets = CosetTable.coset_table_max_limit
         if strategy == 'relator_based':
-            C = coset_enumeration_r(self, H, max_cosets=max_cosets)
+            C = coset_enumeration_r(self, H, max_cosets=max_cosets,
+                                                    resume=resume, incomplete=incomplete)
         else:
-            C = coset_enumeration_c(self, H, max_cosets=max_cosets)
-        C.compress()
+            C = coset_enumeration_c(self, H, max_cosets=max_cosets,
+                                                    resume=resume, incomplete=incomplete)
+        if C.is_complete():
+            C.compress()
         return C
 
     def standardize_coset_table(self):
@@ -174,6 +187,11 @@ class FpGroup(DefaultPrinting):
         and `None` otherwise
 
         '''
+        used_gens = set()
+        for r in self.relators:
+            used_gens.update(r.contains_generators())
+        if any([g not in used_gens for g in self.generators]):
+            return True
         # Abelianisation test: check is the abelianisation is infinite
         abelian_rels = []
         from sympy.polys.solvers import RawMatrix as Matrix
@@ -200,13 +218,16 @@ class FpGroup(DefaultPrinting):
         rels = list(self.generators)
         rels.extend(self.relators)
         if not s:
-            rand = self.free_group.identity
-            i = 0
-            while (rand in rels or rand**-1 in rels or rand.is_identity
-                   or rand in rels) and i<10:
-                rand = self.random_element()
-                i += 1
-            s = [gen, rand] + [g for g in self.generators if g != gen]
+            if len(self.generators) == 2:
+                s = [gen] + [g for g in self.generators if g != gen]
+            else:
+                rand = self.free_group.identity
+                i = 0
+                while ((rand in rels or rand**-1 in rels or rand.is_identity)
+                        and i<10):
+                    rand = self.random_element()
+                    i += 1
+                s = [gen, rand] + [g for g in self.generators if g != gen]
         mid = (len(s)+1)//2
         half1 = s[:mid]
         half2 = s[mid:]
@@ -228,6 +249,7 @@ class FpGroup(DefaultPrinting):
                     continue
         if not C:
             return None, None
+        C.compress()
         return half, C
 
     def most_frequent_generator(self):
@@ -275,984 +297,6 @@ class FpGroup(DefaultPrinting):
         return str_form
 
     __repr__ = __str__
-
-
-###############################################################################
-#                           COSET TABLE                                       #
-###############################################################################
-
-class CosetTable(DefaultPrinting):
-    # coset_table: Mathematically a coset table
-    #               represented using a list of lists
-    # alpha: Mathematically a coset (precisely, a live coset)
-    #       represented by an integer between i with 1 <= i <= n
-    #       α ∈ c
-    # x: Mathematically an element of "A" (set of generators and
-    #   their inverses), represented using "FpGroupElement"
-    # fp_grp: Finitely Presented Group with < X|R > as presentation.
-    # H: subgroup of fp_grp.
-    # NOTE: We start with H as being only a list of words in generators
-    #       of "fp_grp". Since `.subgroup` method has not been implemented.
-
-    r"""
-
-    Properties
-    ==========
-
-    [1] `0 \in \Omega` and `\tau(1) = \epsilon`
-    [2] `\alpha^x = \beta \Leftrightarrow \beta^{x^{-1}} = \alpha`
-    [3] If `\alpha^x = \beta`, then `H \tau(\alpha)x = H \tau(\beta)`
-    [4] `\forall \alpha \in \Omega, 1^{\tau(\alpha)} = \alpha`
-
-    References
-    ==========
-
-    [1] Holt, D., Eick, B., O'Brien, E.
-    "Handbook of Computational Group Theory"
-
-    [2] John J. Cannon; Lucien A. Dimino; George Havas; Jane M. Watson
-    Mathematics of Computation, Vol. 27, No. 123. (Jul., 1973), pp. 463-490.
-    "Implementation and Analysis of the Todd-Coxeter Algorithm"
-
-    """
-    # default limit for the number of cosets allowed in a
-    # coset enumeration.
-    coset_table_max_limit = 4096000
-    # maximum size of deduction stack above or equal to
-    # which it is emptied
-    max_stack_size = 500
-
-    def __init__(self, fp_grp, subgroup, max_cosets=None):
-        if not max_cosets:
-            max_cosets = CosetTable.coset_table_max_limit
-        self.fp_group = fp_grp
-        self.subgroup = subgroup
-        self.coset_table_max_limit = max_cosets
-        # "p" is setup independent of Ω and n
-        self.p = [0]
-        # a list of the form `[gen_1, gen_1^{-1}, ... , gen_k, gen_k^{-1}]`
-        self.A = list(chain.from_iterable((gen, gen**-1) \
-                for gen in self.fp_group.generators))
-        # the mathematical coset table which is a list of lists
-        self.table = [[None]*len(self.A)]
-        self.A_dict = {x: self.A.index(x) for x in self.A}
-        self.A_dict_inv = {}
-        for x, index in self.A_dict.items():
-            if index % 2 == 0:
-                self.A_dict_inv[x] = self.A_dict[x] + 1
-            else:
-                self.A_dict_inv[x] = self.A_dict[x] - 1
-        # used in the coset-table based method of coset enumeration. Each of
-        # the element is called a "deduction" which is the form (α, x) whenever
-        # a value is assigned to α^x during a definition or "deduction process"
-        self.deduction_stack = []
-
-    @property
-    def omega(self):
-        """Set of live cosets. """
-        return [coset for coset in range(len(self.p)) if self.p[coset] == coset]
-
-    def copy(self):
-        """
-        Return a shallow copy of Coset Table instance ``self``.
-
-        """
-        self_copy = self.__class__(self.fp_group, self.subgroup)
-        self_copy.table = [list(perm_rep) for perm_rep in self.table]
-        self_copy.p = list(self.p)
-        self_copy.deduction_stack = list(self.deduction_stack)
-        return self_copy
-
-    def __str__(self):
-        return "Coset Table on %s with %s as subgroup generators" \
-                % (self.fp_group, self.subgroup)
-
-    __repr__ = __str__
-
-    @property
-    def n(self):
-        """The number `n` represents the length of the sublist containing the
-        live cosets.
-
-        """
-        if not self.table:
-            return 0
-        return max(self.omega) + 1
-
-    # Pg. 152 [1]
-    def is_complete(self):
-        r"""
-        The coset table is called complete if it has no undefined entries
-        on the live cosets; that is, `\alpha^x` is defined for all
-        `\alpha \in \Omega` and `x \in A`.
-
-        """
-        return not any([None in self.table[coset] for coset in self.omega])
-
-    # Pg. 153 [1]
-    def define(self, alpha, x, max_cosets=coset_table_max_limit):
-        r"""
-        This routine is used in the relator-based strategy of Todd-Coxeter
-        algorithm if some `\alpha^x` is undefined. We check whether there is
-        space available for defining a new coset. If there is enough space
-        then we remedy this by adjoining a new coset `\beta` to `\Omega`
-        (i.e to set of live cosets) and put that equal to `\alpha^x`, then
-        make an assignment satisfying Property[1]. If there is not enough space
-        then we halt the Coset Table creation. The maximum amount of space that
-        can be used by Coset Table can be manipulated using the class variable
-        ``CosetTable.coset_table_max_limit``.
-
-        See Also
-        ========
-        define_c
-
-        """
-        A = self.A
-        table = self.table
-        len_table = len(table)
-        if len_table == self.coset_table_max_limit:
-            # abort the further generation of cosets
-            raise ValueError("the coset enumeration has defined more than "
-                    "%s cosets. Try with a greater value max number of cosets "
-                    % self.coset_table_max_limit)
-        table.append([None]*len(A))
-        # beta is the new coset generated
-        beta = len_table
-        self.p.append(beta)
-        table[alpha][self.A_dict[x]] = beta
-        table[beta][self.A_dict_inv[x]] = alpha
-
-    def define_c(self, alpha, x, max_cosets=coset_table_max_limit):
-        r"""
-        A variation of ``define`` routine, described on Pg. 165 [1], used in
-        the coset table-based strategy of Todd-Coxeter algorithm. It differs
-        from ``define`` routine in that for each definition it also adds the
-        tuple `(\alpha, x)` to the deduction stack.
-
-        See Also
-        ========
-        define
-
-        """
-        A = self.A
-        table = self.table
-        len_table = len(table)
-        if len_table == self.coset_table_max_limit:
-            # abort the further generation of cosets
-            raise ValueError("the coset enumeration has defined more than "
-                    "%s cosets. Try with a greater value max number of cosets "
-                    % max_cosets)
-        table.append([None]*len(A))
-        # beta is the new coset generated
-        beta = len_table
-        self.p.append(beta)
-        table[alpha][self.A_dict[x]] = beta
-        table[beta][self.A_dict_inv[x]] = alpha
-        # append to deduction stack
-        self.deduction_stack.append((alpha, x))
-
-    def scan_c(self, alpha, word):
-        """
-        A variation of ``scan`` routine, described on pg. 165 of [1], which
-        puts at tuple, whenever a deduction occurs, to deduction stack.
-
-        See Also
-        ========
-        scan, scan_check, scan_and_fill, scan_and_fill_c
-
-        """
-        # α is an integer representing a "coset"
-        # since scanning can be in two cases
-        # 1. for α=0 and w in Y (i.e generating set of H)
-        # 2. α in Ω (set of live cosets), w in R (relators)
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        f = alpha
-        i = 0
-        r = len(word)
-        b = alpha
-        j = r - 1
-        # list of union of generators and their inverses
-        while i <= j and table[f][A_dict[word[i]]] is not None:
-            f = table[f][A_dict[word[i]]]
-            i += 1
-        if i > j:
-            if f != b:
-                self.coincidence_c(f, b)
-            return
-        while j >= i and table[b][A_dict_inv[word[j]]] is not None:
-            b = table[b][A_dict_inv[word[j]]]
-            j -= 1
-        if j < i:
-            # we have an incorrect completed scan with coincidence f ~ b
-            # run the "coincidence" routine
-            self.coincidence_c(f, b)
-        elif j == i:
-            # deduction process
-            table[f][A_dict[word[i]]] = b
-            table[b][A_dict_inv[word[i]]] = f
-            self.deduction_stack.append((f, word[i]))
-        # otherwise scan is incomplete and yields no information
-
-    # α, β coincide, i.e. α, β represent the pair of cosets where
-    # coincidence occurs
-    def coincidence_c(self, alpha, beta):
-        """
-        A variation of ``coincidence`` routine used in the coset-table based
-        method of coset enumeration. The only difference being on addition of
-        a new coset in coset table(i.e new coset introduction), then it is
-        appended to ``deduction_stack``.
-
-        See Also
-        ========
-        coincidence
-
-        """
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        p = self.p
-        l = 0
-        # behaves as a queue
-        q = []
-        self.merge(alpha, beta, q)
-        while len(q) > 0:
-            gamma = q.pop(0)
-            for x in A_dict:
-                delta = table[gamma][A_dict[x]]
-                if delta is not None:
-                    table[delta][A_dict_inv[x]] = None
-                    # only line of difference from ``coincidence`` routine
-                    self.deduction_stack.append((delta, x**-1))
-                    mu = self.rep(gamma)
-                    nu = self.rep(delta)
-                    if table[mu][A_dict[x]] is not None:
-                        self.merge(nu, table[mu][A_dict[x]], q)
-                    elif table[nu][A_dict_inv[x]] is not None:
-                        self.merge(mu, table[nu][A_dict_inv[x]], q)
-                    else:
-                        table[mu][A_dict[x]] = nu
-                        table[nu][A_dict_inv[x]] = mu
-
-    def scan(self, alpha, word):
-        r"""
-        ``scan`` performs a scanning process on the input ``word``.
-        It first locates the largest prefix ``s`` of ``word`` for which
-        `\alpha^s` is defined (i.e is not ``None``), ``s`` may be empty. Let
-        ``word=sv``, let ``t`` be the longest suffix of ``v`` for which
-        `\alpha^{t^{-1}}` is defined, and let ``v=ut``. Then three
-        possibilities are there:
-
-        1. If ``t=v``, then we say that the scan completes, and if, in addition
-        `\alpha^s = \alpha^{t^{-1}}`, then we say that the scan completes
-        correctly.
-
-        2. It can also happen that scan does not complete, but `|u|=1`; that
-        is, the word ``u`` consists of a single generator `x \in A`. In that
-        case, if `\alpha^s = \beta` and `\alpha^{t^{-1}} = \gamma`, then we can
-        set `\beta^x = \gamma` and `\gamma^{x^{-1}} = \beta`. These assignments
-        are known as deductions and enable the scan to complete correctly.
-
-        3. See ``coicidence`` routine for explanation of third condition.
-
-        Notes
-        =====
-        The code for the procedure of scanning `\alpha \in \Omega`
-        under `w \in A*` is defined on pg. 155 [1]
-
-        See Also
-        ========
-        scan_c, scan_check, scan_and_fill, scan_and_fill_c
-
-        """
-        # α is an integer representing a "coset"
-        # since scanning can be in two cases
-        # 1. for α=0 and w in Y (i.e generating set of H)
-        # 2. α in Ω (set of live cosets), w in R (relators)
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        f = alpha
-        i = 0
-        r = len(word)
-        b = alpha
-        j = r - 1
-        while i <= j and table[f][A_dict[word[i]]] is not None:
-            f = table[f][A_dict[word[i]]]
-            i += 1
-        if i > j:
-            if f != b:
-                self.coincidence(f, b)
-            return
-        while j >= i and table[b][A_dict_inv[word[j]]] is not None:
-            b = table[b][A_dict_inv[word[j]]]
-            j -= 1
-        if j < i:
-            # we have an incorrect completed scan with coincidence f ~ b
-            # run the "coincidence" routine
-            self.coincidence(f, b)
-        elif j == i:
-            # deduction process
-            table[f][A_dict[word[i]]] = b
-            table[b][A_dict_inv[word[i]]] = f
-        # otherwise scan is incomplete and yields no information
-
-    # used in the low-index subgroups algorithm
-    def scan_check(self, alpha, word):
-        r"""
-        Another version of ``scan`` routine, described on, it checks whether
-        `\alpha` scans correctly under `word`, it is a straightforward
-        modification of ``scan``. ``scan_check`` returns ``False`` (rather than
-        calling ``coincidence``) if the scan completes incorrectly; otherwise
-        it returns ``True``.
-
-        See Also
-        ========
-        scan, scan_c, scan_and_fill, scan_and_fill_c
-
-        """
-        # α is an integer representing a "coset"
-        # since scanning can be in two cases
-        # 1. for α=0 and w in Y (i.e generating set of H)
-        # 2. α in Ω (set of live cosets), w in R (relators)
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        f = alpha
-        i = 0
-        r = len(word)
-        b = alpha
-        j = r - 1
-        while i <= j and table[f][A_dict[word[i]]] is not None:
-            f = table[f][A_dict[word[i]]]
-            i += 1
-        if i > j:
-            return f == b
-        while j >= i and table[b][A_dict_inv[word[j]]] is not None:
-            b = table[b][A_dict_inv[word[j]]]
-            j -= 1
-        if j < i:
-            # we have an incorrect completed scan with coincidence f ~ b
-            # return False, instead of calling coincidence routine
-            return False
-        elif j == i:
-            # deduction process
-            table[f][A_dict[word[i]]] = b
-            table[b][A_dict_inv[word[i]]] = f
-        return True
-
-    def merge(self, k, lamda, q):
-        """
-        Input: 'k', 'lamda' being the two class representatives to be merged.
-        =====
-
-        Merge two classes with representatives ``k`` and ``lamda``, described
-        on Pg. 157 [1] (for pseudocode), start by putting ``p[k] = lamda``.
-        It is more efficient to choose the new representative from the larger
-        of the two classes being merged, i.e larger among ``k`` and ``lamda``.
-        procedure ``merge`` performs the merging operation, adds the deleted
-        class representative to the queue ``q``.
-
-        Notes
-        =====
-        Pg. 86-87 [1] contains a description of this method.
-
-        See Also
-        ========
-        coincidence, rep
-
-        """
-        p = self.p
-        phi = self.rep(k)
-        psi = self.rep(lamda)
-        if phi != psi:
-            mu = min(phi, psi)
-            v = max(phi, psi)
-            p[v] = mu
-            q.append(v)
-
-    def rep(self, k):
-        r"""
-        Input: `k \in [0 \ldots n-1]`, as for ``self`` only array ``p`` is used
-        =====
-        Output: Representative of the class containing ``k``.
-        ======
-
-        Returns the representative of `\sim` class containing ``k``, it also
-        makes some modification to array ``p`` of ``self`` to ease further
-        computations, described on Pg. 157 [1].
-
-        The information on classes under `\sim` is stored in array `p` of
-        ``self`` argument, which will always satisfy the property:
-
-        `p[\alpha] \sim \alpha` and `p[\alpha]=\alpha \iff \alpha=rep(\alpha)`
-        `\forall \in [0 \ldots n-1]`.
-
-        So, for `\alpha \in [0 \ldots n-1]`, we find `rep(self, \alpha)` by
-        continually replacing `\alpha` by `p[\alpha]` until it becomes
-        constant (i.e satisfies `p[\alpha] = \alpha`):w
-
-        To increase the efficiency of later ``rep`` calculations, whenever we
-        find `rep(self, \alpha)=\beta`, we set
-        `p[\gamma] = \beta \forall \gamma \in p-chain` from `\alpha` to `\beta`
-
-        Notes
-        =====
-        ``rep`` routine is also described on Pg. 85-87 [1] in Atkinson's
-        algorithm, this results from the fact that ``coincidence`` routine
-        introduces functionality similar to that introduced by the
-        ``minimal_block`` routine on Pg. 85-87 [1].
-
-        See also
-        ========
-        coincidence, merge
-
-        """
-        p = self.p
-        lamda = k
-        rho = p[lamda]
-        while rho != lamda:
-            lamda = rho
-            rho = p[lamda]
-        mu = k
-        rho = p[mu]
-        while rho != lamda:
-            p[mu] = lamda
-            mu = rho
-            rho = p[mu]
-        return lamda
-
-    # α, β coincide, i.e. α, β represent the pair of cosets
-    # where coincidence occurs
-    def coincidence(self, alpha, beta):
-        r"""
-        The third situation described in ``scan`` routine is handled by this
-        routine, described on Pg. 156-161 [1].
-
-        The unfortunate situation when the scan completes but not correctly,
-        then ``coincidence`` routine is run. i.e when for some `i` with
-        `1 \le i \le r+1`, we have `w=st` with `s=x_1*x_2 ... x_{i-1}`,
-        `t=x_i*x_{i+1} ... x_r`, and `\beta = \alpha^s` and
-        `\gamma = \alph^{t-1}` are defined but unequal. This means that
-        `\beta` and `\gamma` represent the same coset of `H` in `G`. Described
-        on Pg. 156 [1]. ``rep``
-
-        See Also
-        ========
-        scan
-
-        """
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        p = self.p
-        l = 0
-        # behaves as a queue
-        q = []
-        self.merge(alpha, beta, q)
-        while len(q) > 0:
-            gamma = q.pop(0)
-            for x in A_dict:
-                delta = table[gamma][A_dict[x]]
-                if delta is not None:
-                    table[delta][A_dict_inv[x]] = None
-                    mu = self.rep(gamma)
-                    nu = self.rep(delta)
-                    if table[mu][A_dict[x]] is not None:
-                        self.merge(nu, table[mu][A_dict[x]], q)
-                    elif table[nu][A_dict_inv[x]] is not None:
-                        self.merge(mu, table[nu][A_dict_inv[x]], q)
-                    else:
-                        table[mu][A_dict[x]] = nu
-                        table[nu][A_dict_inv[x]] = mu
-
-    # method used in the HLT strategy
-    def scan_and_fill(self, alpha, word, max_cosets=coset_table_max_limit):
-        """
-        A modified version of ``scan`` routine used in the relator-based
-        method of coset enumeration, described on pg. 162-163 [1], which
-        follows the idea that whenever the procedure is called and the scan
-        is incomplete then it makes new definitions to enable the scan to
-        complete; i.e it fills in the gaps in the scan of the relator or
-        subgroup generator.
-
-        """
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        r = len(word)
-        f = alpha
-        i = 0
-        b = alpha
-        j = r - 1
-        # loop until it has filled the α row in the table.
-        while True:
-            # do the forward scanning
-            while i <= j and table[f][A_dict[word[i]]] is not None:
-                f = table[f][A_dict[word[i]]]
-                i += 1
-            if i > j:
-                if f != b:
-                    self.coincidence(f, b)
-                return
-            # forward scan was incomplete, scan backwards
-            while j >= i and table[b][A_dict_inv[word[j]]] is not None:
-                b = table[b][A_dict_inv[word[j]]]
-                j -= 1
-            if j < i:
-                self.coincidence(f, b)
-            elif j == i:
-                table[f][A_dict[word[i]]] = b
-                table[b][A_dict_inv[word[i]]] = f
-            else:
-                self.define(f, word[i], max_cosets=max_cosets)
-
-    def scan_and_fill_c(self, alpha, word, max_cosets=coset_table_max_limit):
-        """
-        A modified version of ``scan`` routine, described on Pg. 165 second
-        para. [1], with modification similar to that of ``scan_anf_fill`` the
-        only difference being it calls the coincidence procedure used in the
-        coset-table based method i.e. the routine ``coincidence_c`` is used.
-
-        Also See
-        ========
-        scan, scan_and_fill
-
-        """
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        r = len(word)
-        f = alpha
-        i = 0
-        b = alpha
-        j = r - 1
-        # loop until it has filled the α row in the table.
-        while True:
-            # do the forward scanning
-            while i <= j and table[f][A_dict[word[i]]] is not None:
-                f = table[f][A_dict[word[i]]]
-                i += 1
-            if i > j:
-                if f != b:
-                    self.coincidence_c(f, b)
-                return
-            # forward scan was incomplete, scan backwards
-            while j >= i and table[b][A_dict_inv[word[j]]] is not None:
-                b = table[b][A_dict_inv[word[j]]]
-                j -= 1
-            if j < i:
-                self.coincidence_c(f, b)
-            elif j == i:
-                table[f][A_dict[word[i]]] = b
-                table[b][A_dict_inv[word[i]]] = f
-                self.deduction_stack.append((f, word[i]))
-            else:
-                self.define_c(f, word[i], max_cosets=max_cosets)
-
-    # method used in the HLT strategy
-    def look_ahead(self):
-        """
-        When combined with the HLT method this is known as HLT+Lookahead
-        method of coset enumeration, described on pg. 164 [1]. Whenever
-        ``define`` aborts due to lack of space available this procedure is
-        executed. This routine helps in recovering space resulting from
-        "coincidence" of cosets.
-
-        """
-        R = self.fp_group.relators
-        p = self.p
-        # complete scan all relators under all cosets(obviously live)
-        # without making new definitions
-        for beta in self.omega:
-            for w in R:
-                self.scan(beta, w)
-                if p[beta] < beta:
-                    break
-
-    # Pg. 166
-    def process_deductions(self, R_c_x, R_c_x_inv):
-        """
-        Processes the deductions that have been pushed onto ``deduction_stack``,
-        described on Pg. 166 [1] and is used in coset-table based enumeration.
-
-        See Also
-        ========
-        deduction_stack
-
-        """
-        p = self.p
-        table = self.table
-        while len(self.deduction_stack) > 0:
-            if len(self.deduction_stack) >= CosetTable.max_stack_size:
-                self.look_ahead()
-                del self.deduction_stack[:]
-            else:
-                alpha, x = self.deduction_stack.pop()
-                if p[alpha] == alpha:
-                    for w in R_c_x:
-                        self.scan_c(alpha, w)
-                        if p[alpha] < alpha:
-                            break
-            beta = table[alpha][self.A_dict[x]]
-            if beta is not None and p[beta] == beta:
-                for w in R_c_x_inv:
-                    self.scan_c(beta, w)
-                    if p[beta] < beta:
-                        break
-
-    def process_deductions_check(self, R_c_x, R_c_x_inv):
-        """
-        A variation of ``process_deductions``, this calls ``scan_check``
-        wherever ``process_deductions`` calls ``scan``, described on Pg. [1].
-
-        See Also
-        ========
-        process_deductions
-
-        """
-        p = self.p
-        table = self.table
-        while len(self.deduction_stack) > 0:
-            alpha, x = self.deduction_stack.pop()
-            for w in R_c_x:
-                if not self.scan_check(alpha, w):
-                    return False
-            beta = table[alpha][self.A_dict[x]]
-            if beta is not None:
-                for w in R_c_x_inv:
-                    if not self.scan_check(beta, w):
-                        return False
-        return True
-
-    def switch(self, beta, gamma):
-        r"""Switch the elements `\beta, \gamma \in \Omega` of ``self``, used
-        by the ``standardize`` procedure, described on Pg. 167 [1].
-
-        See Also
-        ========
-        standardize
-
-        """
-        A = self.A
-        A_dict = self.A_dict
-        table = self.table
-        for x in A:
-            z = table[gamma][A_dict[x]]
-            table[gamma][A_dict[x]] = table[beta][A_dict[x]]
-            table[beta][A_dict[x]] = z
-            for alpha in range(len(self.p)):
-                if self.p[alpha] == alpha:
-                    if table[alpha][A_dict[x]] == beta:
-                        table[alpha][A_dict[x]] = gamma
-                    elif table[alpha][A_dict[x]] == gamma:
-                        table[alpha][A_dict[x]] = beta
-
-    def standardize(self):
-        """
-        A coset table is standardized if when running through the cosets and
-        within each coset through the generator images (ignoring generator
-        inverses), the cosets appear in order of the integers
-        `0, 1, , \ldots, n`. "Standardize" reorders the elements of `\Omega`
-        such that, if we scan the coset table first by elements of `\Omega`
-        and then by elements of A, then the cosets occur in ascending order.
-        ``standardize()`` is used at the end of an enumeration to permute the
-        cosets so that they occur in some sort of standard order.
-
-        Notes
-        =====
-        procedure is described on pg. 167-168 [1], it also makes use of the
-        ``switch`` routine to replace by smaller integer value.
-
-        >>> from sympy.combinatorics.free_groups import free_group
-        >>> from sympy.combinatorics.fp_groups import FpGroup, coset_enumeration_r
-        >>> F, x, y = free_group("x, y")
-
-        # Example 5.3 from [1]
-        >>> f = FpGroup(F, [x**2*y**2, x**3*y**5])
-        >>> C = coset_enumeration_r(f, [])
-        >>> C.compress()
-        >>> C.table
-        [[1, 3, 1, 3], [2, 0, 2, 0], [3, 1, 3, 1], [0, 2, 0, 2]]
-        >>> C.standardize()
-        >>> C.table
-        [[1, 2, 1, 2], [3, 0, 3, 0], [0, 3, 0, 3], [2, 1, 2, 1]]
-
-        """
-        A = self.A
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        gamma = 1
-        for alpha, x in product(range(self.n), A):
-            beta = self.table[alpha][A_dict[x]]
-            if beta >= gamma:
-                if beta > gamma:
-                    self.switch(gamma, beta)
-                gamma += 1
-                if gamma == self.n:
-                    return
-
-    # Compression of a Coset Table
-    def compress(self):
-        """Removes the non-live cosets from the coset table, described on
-        pg. 167 [1].
-
-        """
-        gamma = -1
-        A = self.A
-        A_dict = self.A_dict
-        A_dict_inv = self.A_dict_inv
-        table = self.table
-        chi = tuple([i for i in range(len(self.p)) if self.p[i] != i])
-        for alpha in self.omega:
-            gamma += 1
-            if gamma != alpha:
-                # replace α by γ in coset table
-                for x in A:
-                    beta = table[alpha][A_dict[x]]
-                    table[gamma][A_dict[x]] = beta
-                    table[beta][A_dict_inv[x]] == gamma
-        # all the cosets in the table are live cosets
-        self.p = list(range(gamma + 1))
-        # delete the useless coloumns
-        del table[len(self.p):]
-        # re-define values
-        for row in table:
-            for j in range(len(self.A)):
-                row[j] -= bisect_left(chi, row[j])
-
-    def conjugates(self, R):
-        R_c = list(chain.from_iterable((rel.cyclic_conjugates(), \
-                (rel**-1).cyclic_conjugates()) for rel in R))
-        R_set = set()
-        for conjugate in R_c:
-            R_set = R_set.union(conjugate)
-        R_c_list = []
-        for x in self.A:
-            r = set([word for word in R_set if word[0] == x])
-            R_c_list.append(r)
-            R_set.difference_update(r)
-        return R_c_list
-
-###############################################################################
-#                           COSET ENUMERATION                                 #
-###############################################################################
-
-# relator-based method
-def coset_enumeration_r(fp_grp, Y, max_cosets=None):
-    """
-    This is easier of the two implemented methods of coset enumeration.
-    and is often called the HLT method, after Hazelgrove, Leech, Trotter
-    The idea is that we make use of ``scan_and_fill`` makes new definitions
-    whenever the scan is incomplete to enable the scan to complete; this way
-    we fill in the gaps in the scan of the relator or subgroup generator,
-    that's why the name relator-based method.
-
-    # TODO: complete the docstring
-
-    See Also
-    ========
-    scan_and_fill,
-
-    References
-    ==========
-
-    [1] Holt, D., Eick, B., O'Brien, E.
-    "Handbook of computational group theory"
-
-    >>> from sympy.combinatorics.free_groups import free_group
-    >>> from sympy.combinatorics.fp_groups import FpGroup, coset_enumeration_r
-    >>> F, x, y = free_group("x, y")
-
-    # Example 5.1 from [1]
-    >>> f = FpGroup(F, [x**3, y**3, x**-1*y**-1*x*y])
-    >>> C = coset_enumeration_r(f, [x])
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         print(C.table[i])
-    [0, 0, 1, 2]
-    [1, 1, 2, 0]
-    [2, 2, 0, 1]
-    >>> C.p
-    [0, 1, 2, 1, 1]
-
-    # Example from exercises Q2 [1]
-    >>> f = FpGroup(F, [x**2*y**2, y**-1*x*y*x**-3])
-    >>> C = coset_enumeration_r(f, [])
-    >>> C.compress(); C.standardize()
-    >>> C.table
-    [[1, 2, 3, 4],
-    [5, 0, 6, 7],
-    [0, 5, 7, 6],
-    [7, 6, 5, 0],
-    [6, 7, 0, 5],
-    [2, 1, 4, 3],
-    [3, 4, 2, 1],
-    [4, 3, 1, 2]]
-
-    # Example 5.2
-    >>> f = FpGroup(F, [x**2, y**3, (x*y)**3])
-    >>> Y = [x*y]
-    >>> C = coset_enumeration_r(f, Y)
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         print(C.table[i])
-    [1, 1, 2, 1]
-    [0, 0, 0, 2]
-    [3, 3, 1, 0]
-    [2, 2, 3, 3]
-
-    # Example 5.3
-    >>> f = FpGroup(F, [x**2*y**2, x**3*y**5])
-    >>> Y = []
-    >>> C = coset_enumeration_r(f, Y)
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         print(C.table[i])
-    [1, 3, 1, 3]
-    [2, 0, 2, 0]
-    [3, 1, 3, 1]
-    [0, 2, 0, 2]
-
-    # Example 5.4
-    >>> F, a, b, c, d, e = free_group("a, b, c, d, e")
-    >>> f = FpGroup(F, [a*b*c**-1, b*c*d**-1, c*d*e**-1, d*e*a**-1, e*a*b**-1])
-    >>> Y = [a]
-    >>> C = coset_enumeration_r(f, Y)
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         print(C.table[i])
-    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-
-    # example of "compress" method
-    >>> C.compress()
-    >>> C.table
-    [[0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
-
-    # Exercises Pg. 161, Q2.
-    >>> F, x, y = free_group("x, y")
-    >>> f = FpGroup(F, [x**2*y**2, y**-1*x*y*x**-3])
-    >>> Y = []
-    >>> C = coset_enumeration_r(f, Y)
-    >>> C.compress()
-    >>> C.standardize()
-    >>> C.table
-    [[1, 2, 3, 4],
-    [5, 0, 6, 7],
-    [0, 5, 7, 6],
-    [7, 6, 5, 0],
-    [6, 7, 0, 5],
-    [2, 1, 4, 3],
-    [3, 4, 2, 1],
-    [4, 3, 1, 2]]
-
-    # John J. Cannon; Lucien A. Dimino; George Havas; Jane M. Watson
-    # Mathematics of Computation, Vol. 27, No. 123. (Jul., 1973), pp. 463-490
-    # from 1973chwd.pdf
-    # Table 1. Ex. 1
-    >>> F, r, s, t = free_group("r, s, t")
-    >>> E1 = FpGroup(F, [t**-1*r*t*r**-2, r**-1*s*r*s**-2, s**-1*t*s*t**-2])
-    >>> C = coset_enumeration_r(E1, [r])
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         print(C.table[i])
-    [0, 0, 0, 0, 0, 0]
-
-    Ex. 2
-    >>> F, a, b = free_group("a, b")
-    >>> Cox = FpGroup(F, [a**6, b**6, (a*b)**2, (a**2*b**2)**2, (a**3*b**3)**5])
-    >>> C = coset_enumeration_r(Cox, [a])
-    >>> index = 0
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         index += 1
-    >>> index
-    500
-
-    # Ex. 3
-    >>> F, a, b = free_group("a, b")
-    >>> B_2_4 = FpGroup(F, [a**4, b**4, (a*b)**4, (a**-1*b)**4, (a**2*b)**4, \
-            (a*b**2)**4, (a**2*b**2)**4, (a**-1*b*a*b)**4, (a*b**-1*a*b)**4])
-    >>> C = coset_enumeration_r(B_2_4, [a])
-    >>> index = 0
-    >>> for i in range(len(C.p)):
-    ...     if C.p[i] == i:
-    ...         index += 1
-    >>> index
-    1024
-
-    """
-    # 1. Initialize a coset table C for < X|R >
-    C = CosetTable(fp_grp, Y, max_cosets=max_cosets)
-    R = fp_grp.relators
-    A_dict = C.A_dict
-    A_dict_inv = C.A_dict_inv
-    p = C.p
-    for w in Y:
-        C.scan_and_fill(0, w)
-    alpha = 0
-    while alpha < C.n:
-        if p[alpha] == alpha:
-            for w in R:
-                C.scan_and_fill(alpha, w)
-                # if α was eliminated during the scan then break
-                if p[alpha] < alpha:
-                    break
-            if p[alpha] == alpha:
-                for x in A_dict:
-                    if C.table[alpha][A_dict[x]] is None:
-                        C.define(alpha, x)
-        alpha += 1
-    return C
-
-
-# Pg. 166
-# coset-table based method
-def coset_enumeration_c(fp_grp, Y, max_cosets=None):
-    """
-    >>> from sympy.combinatorics.free_groups import free_group
-    >>> from sympy.combinatorics.fp_groups import FpGroup, coset_enumeration_c
-    >>> F, x, y = free_group("x, y")
-    >>> f = FpGroup(F, [x**3, y**3, x**-1*y**-1*x*y])
-    >>> C = coset_enumeration_c(f, [x])
-    >>> C.table
-    [[0, 0, 1, 2], [1, 1, 2, 0], [2, 2, 0, 1]]
-
-    """
-    # Initialize a coset table C for < X|R >
-    C = CosetTable(fp_grp, Y)
-    X = fp_grp.generators
-    R = fp_grp.relators
-    A = C.A
-    # replace all the elements by cyclic reductions
-    R_cyc_red = [rel.identity_cyclic_reduction() for rel in R]
-    R_c = list(chain.from_iterable((rel.cyclic_conjugates(), (rel**-1).cyclic_conjugates()) \
-            for rel in R_cyc_red))
-    R_set = set()
-    for conjugate in R_c:
-        R_set = R_set.union(conjugate)
-    # a list of subsets of R_c whose words start with "x".
-    R_c_list = []
-    for x in C.A:
-        r = set([word for word in R_set if word[0] == x])
-        R_c_list.append(r)
-        R_set.difference_update(r)
-    for w in Y:
-        C.scan_and_fill_c(0, w)
-    for x in A:
-        C.process_deductions(R_c_list[C.A_dict[x]], R_c_list[C.A_dict_inv[x]])
-    alpha = 0
-    while alpha < len(C.table):
-        if C.p[alpha] == alpha:
-            for x in C.A:
-                if C.p[alpha] != alpha:
-                    break
-                if C.table[alpha][C.A_dict[x]] is None:
-                    C.define_c(alpha, x)
-                    C.process_deductions(R_c_list[C.A_dict[x]], R_c_list[C.A_dict_inv[x]])
-        alpha += 1
-    return C
 
 
 ###############################################################################
@@ -1529,11 +573,11 @@ def reidemeister_relators(C):
     # replace order 1 generators by identity element in reidemeister relators
     for i in range(len(rels)):
         w = rels[i]
-        for gen in order_1_gens:
-            w = w.eliminate_word(gen, identity)
+        w = w.eliminate_words(order_1_gens, _all=True)
         rels[i] = w
 
-    C._schreier_generators = [i for i in C._schreier_generators if i not in order_1_gens]
+    C._schreier_generators = [i for i in C._schreier_generators
+                    if not (i in order_1_gens or i**-1 in order_1_gens)]
 
     # Tietze transformation 1 i.e TT_1
     # remove cyclic conjugate elements from relators
@@ -1614,7 +658,7 @@ def elimination_technique_1(C):
                 gen_index = rel.index(gen**k)
                 bk = rel.subword(gen_index + 1, len(rel))
                 fw = rel.subword(0, gen_index)
-                chi = (bk*fw).identity_cyclic_reduction()
+                chi = bk*fw
                 redundant_gens[gen] = chi**(-1*k)
                 used_gens.update(chi.contains_generators())
                 redundant_rels.append(rel)
@@ -1805,12 +849,11 @@ def reidemeister_presentation(fp_grp, H, C=None):
     g = free_group(syms)[0]
     subs = dict(zip(syms,g.generators))
     C._schreier_generators = g.generators
-    for j in range(len(C._reidemeister_relators)):
-        r = C._reidemeister_relators[j]
+    for j, r in enumerate(C._reidemeister_relators):
         a = r.array_form
         rel = g.identity
-        for i in range(len(a)):
-            rel = rel*subs[a[i][0]]**a[i][1]
+        for sym, p in a:
+            rel = rel*subs[sym]**p
         C._reidemeister_relators[j] = rel
 
     C.schreier_generators = tuple(C._schreier_generators)
