+++ sympy/physics/units/tests/test_quantities.py
@@ -2,15 +2,17 @@
 
 from __future__ import division
 
-from sympy import Symbol, Add, Number, S, integrate, sqrt, Rational, Abs, diff, symbols, Basic
-from sympy.physics.units import convert_to, find_unit
-
-from sympy.physics.units.definitions import s, m, kg, speed_of_light, day, minute, km, foot, meter, grams, amu, au, \
-    quart, inch, coulomb, millimeter, steradian, second, mile, centimeter, hour
-from sympy.physics.units.dimensions import length, time, charge
-from sympy.physics.units.quantities import Quantity
+from sympy import Abs, Add, Basic, Number, Rational, S, Symbol, diff, exp, \
+    integrate, log, sqrt, symbols
+from sympy.physics.units import convert_to, find_unit, amount_of_substance, \
+    volume
+from sympy.physics.units.definitions import amu, au, centimeter, coulomb, day, \
+    foot, grams, hour, inch, kg, km, m, meter, mile, millimeter, minute, mole, \
+    quart, s, second, speed_of_light, steradian
+from sympy.physics.units.dimensions import Dimension, charge, length, time
 from sympy.physics.units.prefixes import PREFIXES, kilo
-from sympy.utilities.pytest import raises
+from sympy.physics.units.quantities import Quantity
+from sympy.utilities.pytest import XFAIL, raises
 
 k = PREFIXES["k"]
 
@@ -236,3 +238,37 @@ def test_sum_of_incompatible_quantities():
     assert expr in Basic._constructor_postprocessor_mapping
     for i in expr.args:
         assert i in Basic._constructor_postprocessor_mapping
+
+
+def test_factor_and_dimension():
+    assert (3000, Dimension(1)) == Quantity._collect_factor_and_dimension(3000)
+    assert (1001, length) == Quantity._collect_factor_and_dimension(meter + km)
+    assert (2, length/time) == Quantity._collect_factor_and_dimension(
+        meter/second + 36*km/(10*hour))
+
+    x, y = symbols('x y')
+    assert (x + y/100, length) == Quantity._collect_factor_and_dimension(
+        x*m + y*centimeter)
+
+    cH = Quantity('cH', amount_of_substance/volume)
+    pH = -log(cH)
+
+    assert (1, volume/amount_of_substance) == Quantity._collect_factor_and_dimension(
+        exp(pH))
+
+    v_w1 = Quantity('v_w1', length/time, S(3)/2*meter/second)
+    v_w2 = Quantity('v_w2', length/time, 2*meter/second)
+    expr = Abs(v_w1/2 - v_w2)
+    assert (S(5)/4, length/time) == \
+        Quantity._collect_factor_and_dimension(expr)
+
+    expr = S(5)/2*second/meter*v_w1 - 3000
+    assert (-(2996 + S(1)/4), Dimension(1)) == \
+        Quantity._collect_factor_and_dimension(expr)
+
+
+@XFAIL
+def test_factor_and_dimension_with_Abs():
+    v_w1 = Quantity('v_w1', length/time, S(3)/2*meter/second)
+    expr = v_w1 - Abs(v_w1)
+    assert (0, lenth/time) == Quantity._collect_factor_and_dimension(expr)
