+++ sympy/release.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.1.dev"
+__version__ = "1.1.2.dev"
