+++ sympy/utilities/tests/test_wester.py
