+++ sympy/vector/coordsysrect.py
@@ -7,7 +7,6 @@ from sympy.solvers import solve
 from sympy.vector.scalar import BaseScalar
 from sympy import eye, trigsimp, ImmutableMatrix as Matrix, Symbol, sin, cos, sqrt, diff, Tuple, acos, atan2, simplify
 import sympy.vector
-from sympy import simplify
 from sympy.vector.orienters import (Orienter, AxisOrienter, BodyOrienter,
                                     SpaceOrienter, QuaternionOrienter)
 
@@ -204,33 +203,36 @@ class CoordSys3D(Basic):
         if isinstance(curv_coord_type, string_types):
             self._set_transformation_equations_mapping(curv_coord_type)
             self._set_lame_coefficient_mapping(curv_coord_type)
+
             if inverse:
                 self._set_inv_trans_equations(curv_coord_type)
 
-        elif isinstance(curv_coord_type, (tuple, list, Tuple)) and len(curv_coord_type) == 3:
-            if all(i.atoms(BaseScalar) for i in curv_coord_type):
+        elif isinstance(curv_coord_type, (tuple, list, Tuple)):
+            if len(curv_coord_type) == 3 and all(i.atoms(BaseScalar) for i in curv_coord_type):
                 self._transformation_eqs = curv_coord_type
-                self._h1, self._h2, self._h3 = self._calculate_lame_coefficients(curv_coord_type)
-                if inverse:
-                    self._inv_transformation_eqs = self._calculate_inv_transformation_equations(
-                                                    self._transformation_equations())
+                self._h1, self._h2, self._h3 = \
+                    self._calculate_lame_coefficients(curv_coord_type)
+            elif len(curv_coord_type) == 2:
+                self._transformation_eqs = \
+                    tuple([eq.subs({curv_coord_type[0][0]: self.x,
+                                    curv_coord_type[0][1]: self.y,
+                                    curv_coord_type[0][2]: self.z})
+                           for eq in curv_coord_type[1]])
+                self._h1, self._h2, self._h3 = \
+                    self._calculate_lame_coefficients(self._transformation_equations())
             else:
                 raise ValueError("Wrong set of parameter.")
 
-        elif isinstance(curv_coord_type, (tuple, list, Tuple)) and len(curv_coord_type) == 2:
-            self._transformation_eqs = \
-            tuple([eq.subs({curv_coord_type[0][0]: self.x,
-                            curv_coord_type[0][1]: self.y,
-                            curv_coord_type[0][2]: self.z}) for eq in curv_coord_type[1]])
-            self._h1, self._h2, self._h3 = self._calculate_lame_coefficients(self._transformation_equations())
             if inverse:
                 self._inv_transformation_eqs = self._calculate_inv_transformation_equations(
-                                                self._transformation_equations())
+                    self._transformation_equations())
+
         else:
             raise ValueError("Wrong set of parameter.")
 
         if not self._check_orthogonality():
-            raise ValueError("The transformation equation does not create orthogonal coordinate system")
+            raise ValueError("The transformation equation does not "
+                             "create orthogonal coordinate system")
 
     def _check_orthogonality(self):
         """
@@ -252,7 +254,7 @@ class CoordSys3D(Basic):
         v2 = Matrix([diff(eq[0], self.y), diff(eq[1], self.y), diff(eq[2], self.y)])
         v3 = Matrix([diff(eq[0], self.z), diff(eq[1], self.z), diff(eq[2], self.z)])
 
-        if any(simplify(i[0]+i[1]+i[2]) == 0 for i in (v1, v2, v3)):
+        if any(simplify(i[0] + i[1] + i[2]) == 0 for i in (v1, v2, v3)):
             return False
         else:
             if simplify(v1.dot(v2)) == 0 and simplify(v2.dot(v3)) == 0 and simplify(v3.dot(v1)) == 0:
@@ -301,7 +303,7 @@ class CoordSys3D(Basic):
 
         equations_mapping = {
             'cartesian': (self.x, self.y, self.z),
-            'spherical': (sqrt(self.x**2 + self.y**2 +self.z**2),
+            'spherical': (sqrt(self.x**2 + self.y**2 + self.z**2),
                           acos((self.z) / sqrt(self.x**2 + self.y**2 + self.z**2)),
                           atan2(self.y, self.x)),
             'cylindrical': (sqrt(self.x**2 + self.y**2),
@@ -333,7 +335,8 @@ class CoordSys3D(Basic):
             'cylindrical': (1, self.y, 1)
         }
         if curv_coord_name not in coefficient_mapping:
-            raise ValueError('Wrong set of parameters. Type of coordinate system is defined')
+            raise ValueError('Wrong set of parameters.'
+                             ' Type of coordinate system is not defined')
         self._h1, self._h2, self._h3 = coefficient_mapping[curv_coord_name]
 
     def _calculate_inv_transformation_equations(self, equations):
